+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/DefaultFilteredTableCellRenderer.java
@@ -82,6 +82,11 @@ final class DefaultFilteredTableCellRenderer<R, C> extends DefaultTableCellRende
   }
 
   @Override
+  public int horizontalAlignment() {
+    return getHorizontalAlignment();
+  }
+
+  @Override
   public Component getTableCellRendererComponent(JTable table, Object value, boolean isSelected,
                                                  boolean hasFocus, int row, int column) {
     super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
@@ -163,6 +168,11 @@ final class DefaultFilteredTableCellRenderer<R, C> extends DefaultTableCellRende
     @Override
     public boolean alternateRowColoring() {
       return alternateRowColoring;
+    }
+
+    @Override
+    public int horizontalAlignment() {
+      return getHorizontalAlignment();
     }
 
     @Override